Task_4_version_2

+++ BlogAsp.BLL/Services/ArticleService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Data.Entity.Core;
 using System.Linq;
 using BlogAsp.BLL.DALInterfaces;
@@ -33,11 +34,15 @@ namespace BlogAsp.BLL.Services
 
         public void Create(Article article)
         {
+            article.Date = DateTime.UtcNow;
+
             _unitOfWork.ArticleGenericRepository.Create(article);
         }
 
         public void Update(Article article)
         {
+            article.Date = DateTime.UtcNow;
+
             _unitOfWork.ArticleGenericRepository.Update(article);
         }
 
@@ -46,20 +51,33 @@ namespace BlogAsp.BLL.Services
             return _unitOfWork.ArticleGenericRepository.Get(id).Tags;
         }
 
-        public Article GetTvTariffWithChannels(Article article, string[] names)
+        public void GetArticlesWithTags(Article article, IEnumerable<string> names)
         {
+            var tags = _unitOfWork.TagGenericRepository.Find(tag => names.Contains(tag.Text));
 
-            foreach (var name in names)
-            {
-                article.Tags.Add(
-                    (_unitOfWork.TagGenericRepository.Find(tag => tag.Text == name).FirstOrDefault()));
-            }
+            article.Tags = tags.ToList();
+        }
 
-            return article;
+        public IEnumerable<string> GetMostPopularTags(Article article, int count)
+        {
+            var words = article.Text.Split(' ');
+
+            var mostPopularWords = words.Where(c => c.Length > 5)
+            .GroupBy(c => c).OrderByDescending(g => g.Count())
+            .Take(count);
+
+            var tags = mostPopularWords.Select(item => item.Key);
+
+            return tags;
         }
 
         public void Delete(int id)
         {
+            if (id <= 0)
+            {
+                throw new ObjectNotFoundException(nameof(Article));
+            }
+
             _unitOfWork.ArticleGenericRepository.Delete(id);
         }
     }
